use class name for categories in impact POIs
Becaused we use a hacky subclass to get corresponding icon on impact
POIs they are displayed with weird category names (eg. florist for
circuits courts).

This change uses the class name instead of subclass name for impact POIs
so that the displayed category name is correct while keeping the same
icon.